# openapi.yaml
openapi: 3.0.3
info:
  title: Status Dashboard API
  version: 1.0.0
servers:
  - url: http://localhost:8000
...
paths:
...
  /auth/callback:
    get:
      summary: The callback URL for the keycloak auth realm. It exchanges the code from keycloak and store tokens for a user. It redirects to the frontend callback URL from the state parameter.
      tags:
        - authentication
      parameters:
        - name: code
          in: query
          description: The code from the keycloak auth realm.
          required: true
          schema:
            type: string
            example: 7b3b3b3b-3b3b-3b3b-3b3b-3b3b3b3b3b3b
        - name: state
          in: query
          description: The state from the authentication request.
          required: true
          schema:
            type: string
            example: eyJjb2RlX2NoYWxsZW5nZSI6IjY0Y2MwYWIxYTg4ZWZlYWNkNjRmYTc5ZWNlMzRlZGUwNDRjZDZkMWMzMmMyYTFjMjc5MWU1YmEyMDYzYzFiZWEiLCJjYWxsYmFja191cmwiOiJodHRwOi8vbG9jYWxob3N0OjUxNzMvY2FsbGJhY2sifQ
      responses:
        '303':
          description: The redirect to a frontend callback url.
        '400':
          description: Return this code with a redirect to a frontend url.
          headers:
            Set-Cookie:
              description: Error cookie set by the server.
              schema:
                type: string
                example: error=some_error_message; Path=/; HttpOnly
...